Backend/requirements.txt
fastapi
uvicorn
redis
pydantic
python-multipart
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
click==8.1.7
Deprecated==1.2.14
exceptiongroup==1.2.2
fastapi==0.115.2
h11==0.14.0
idna==3.10
importlib_resources==6.4.5
limits==3.13.0
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-multipart==0.0.12
redis==5.1.1
slowapi==0.1.9
sniffio==1.3.1
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.32.0
wrapt==1.16.0
